updated jwt and role based authentication, now it works

+++ Labs8/src/main/java/com/javatech/labs8/service/AccountService.java
@@ -1,9 +1,13 @@
 package com.javatech.labs8.service;
 
-import com.javatech.labs8.dtos.UserRegisterDTO;
-import com.javatech.labs8.dtos.UserUpdateDTO;
+import com.javatech.labs8.dtos.AccountDTO;
+import com.javatech.labs8.dtos.AccountLoginDTO;
+import com.javatech.labs8.dtos.AccountRegisterDTO;
+import com.javatech.labs8.dtos.AccountUpdateDTO;
 import com.javatech.labs8.entity.Account;
 import com.javatech.labs8.exceptions.AccountAlreadyExistsException;
+import com.javatech.labs8.exceptions.AccountInvalidCredentialsException;
+import com.javatech.labs8.exceptions.AccountNotFoundException;
 import com.javatech.labs8.pemissions.Role;
 import com.javatech.labs8.repository.AccountRepository;
 
@@ -14,21 +18,73 @@ public class AccountService {
     @Inject
     protected AccountRepository accountRepository;
 
-    public void create(UserRegisterDTO user) throws AccountAlreadyExistsException {
+    public void create(AccountRegisterDTO user) throws AccountAlreadyExistsException {
 
-        boolean exists = !accountRepository.checkIfExists(user.getUsername());
+        boolean exists = accountRepository.checkIfExists(user.getUsername());
+
         if(!exists) {
             accountRepository.save(new Account(user.getUsername(),user.getPassword(), Role.AUTHOR.toString()));
         } else {
             throw new AccountAlreadyExistsException();
         }
 
+
     }
-    public void delete(Long id) {
-        accountRepository.deleteById(Account.class,id);
+    public void delete(Long id) throws AccountNotFoundException {
+
+        boolean exists = accountRepository.checkIfExistsById(id);
+        if(!exists) {
+            accountRepository.deleteById(Account.class,id);
+        } else{
+            throw new AccountNotFoundException();
+        }
+
     }
 
-    public void update(Long id, UserUpdateDTO user) {
+    public AccountDTO findById(Long id) throws AccountNotFoundException {
+        boolean exists = accountRepository.checkIfExistsById(id);
+
+        if(!exists) {
+            Account account = accountRepository.findById(Account.class,id);
+            AccountDTO accountDTO = new AccountDTO();
+
+            accountDTO.setUsername(account.getName());
+            accountDTO.setPassword(account.getPassword());
+            accountDTO.setRole(account.getRole());
+            accountDTO.setId(account.getId());
+
+            return accountDTO;
+        } else {
+            throw new AccountNotFoundException();
+        }
+    }
+
+    public Role getAccountRole(Long id) {
+        boolean exists = accountRepository.checkIfExistsById(id);
+
+        if(exists) {
+            return Role.fromString(accountRepository.findById(Account.class,id).getRole());
+        }
+        return null;
+    }
+
+    public Long validate(AccountLoginDTO user) throws AccountInvalidCredentialsException, AccountNotFoundException {
+        boolean exists = accountRepository.checkIfExists(user.getUsername());
+
+        if(exists) {
+            Account account = accountRepository.getUserByName(user.getUsername());
+
+            if(account.getPassword().equals(user.getPassword())) {
+                return account.getId();
+            } else {
+                throw new AccountInvalidCredentialsException();
+            }
+        } else {
+            throw new AccountNotFoundException();
+        }
+    }
+
+    public void update(Long id, AccountUpdateDTO user) {
 
     }
